+++ feincms3_forms/admin.py
@@ -100,8 +100,15 @@ class SimpleFieldInline(FormFieldInline):
         elif model.TYPE in {T.CHECKBOX}:
             kwargs.setdefault("advanced_fields", ["help_text", "default_value"])
 
+        elif model.TYPE in {T.SELECT}:
+            kwargs.setdefault(
+                "core_fields", ["name", "label", "is_required", "choices"]
+            )
+            kwargs.setdefault(
+                "advanced_fields", ["help_text", "placeholder", "default_value"]
+            )
+
         elif model.TYPE in {
-            T.SELECT,
             T.RADIO,
             T.SELECT_MULTIPLE,
             T.CHECKBOX_SELECT_MULTIPLE,